MDL-57635 tool_usertours: Restrict tours by user account details
This feature adds a filter for determining if a tour can be shown to a user based on when their account
was created, when they first accessed Moodle, when they last logged in to Moodle, or when the tour was
last updated.
This will allow the prevention of tour display to experienced Moodle users, allowing a tour to be
targetted at those who have been using Moodle for a shortened period and presumably require tour
information more.