MDL-66951 tool_licensemanager: Add custom licenses to Moodle
This feature moves the license manager out of the licenselib and makes it pluggable.
Now custom licenses can be added and amended in Moodle, and the site
default can be set to a custom license.

Core licenses remain hard-coded and are uneditable, so they will always
require update within Moodle core updates, and maintain their
internationalisation through core language strings.