bug: file license must remain same even when license disabled
The file license in filepicker and filemanager was reverting to the
site default if the license for a file chosen was disabled.
This behaviour is incorrect and could lead to accidental saving of a
new license for a file that already has a license.
Amended the population of file license list on a case by case basis,
to ensure correct license is always displayed.