bug: prevent deletion of license when being utilised by a file
Custom licenses were able to be deleted even when in use by a file.
This means that we could get into a strange state where we have
a file with a license that no longer exists.
Added logic to check if any files are using a license before deleting
it to prevent this from occurring.