ui: refactor license manager table and edit form for accessibility
Changes include:

- The license manager table was missing any identifiers to assist
screen readers which license rows represent, added columns classes
and row ids to increase ease of access.

- The license editing form was using static html for license short
name when editing, this is counterintuitive for screen readers
looking for form elements as the static form element removes the
label and input tags, convert to a frozen text input instead for
clarity and ease of access.

- chore: convert array syntax to short array (php 5.4+)